[wip] more moving pretty-printing code
