+++ src/swarm-lang/Swarm/Language/Capability.hs
@@ -33,8 +33,10 @@ import Data.Tuple (swap)
 import Data.Yaml
 import GHC.Generics (Generic)
 import Generic.Data (FiniteEnumeration (..))
+import Prettyprinter (pretty)
 import Swarm.Language.Syntax.Constants (Const (..), allConst, constInfo, syntax)
-import Swarm.Util (binTuples, failT)
+import Swarm.Pretty (PrettyPrec (..))
+import Swarm.Util (binTuples, failT, showEnum)
 import Witch (from)
 import Prelude hiding (lookup)
