+++ swarm.cabal
@@ -852,10 +852,12 @@ library swarm-util
     boolexpr,
     clock,
     containers,
+    data-fix,
     directory,
     either,
     extra,
     filepath,
+    free,
     fused-effects,
     hashable,
     lens,
